requirements.txt
altgraph==0.17.4
blinker==1.9.0
click==8.1.7
colorama==0.4.6
easyocr==1.7.2
filelock==3.16.1
Flask==3.1.0
fsspec==2024.10.0
imageio==2.36.1
itsdangerous==2.2.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.2
numpy==2.1.3
opencv-python-headless==4.10.0.84
packaging==24.2
pefile==2023.2.7
pillow==11.0.0
pyclipper==1.3.0.post6
PyGetWindow==0.0.9
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyRect==0.2.0
python-bidi==0.6.3
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
scikit-image==0.24.0
scipy==1.14.1
shapely==2.0.6
sympy==1.13.1
tifffile==2024.9.20
torch==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
Werkzeug==3.1.3
